+++ src/components/AboutMe/AboutMe.css
@@ -1,6 +1,7 @@
 .about-me{
     padding: 110px 70px 15px;
     margin: 0 auto;
+    max-width: 1280px;
 }
 
 .about-me__title{
