+++ src/components/AboutProject/AboutProject.css
@@ -1,6 +1,8 @@
 .about-project{
     padding: 110px 70px;
     margin: 0 auto;
+    max-width: 1280px;
+    width: 100%;
 }
 
 .about-project__subtitle{
