+++ src/components/AboutProject/AboutProject.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './AboutProject.css';
 
 function AboutProject () {
     return(
