+++ src/components/Login/Login.css
@@ -31,7 +31,8 @@
      font-weight: 500;
      font-size: 11px;
      line-height: 13px;
-     width: 396px;
+     max-width: 396px;
+     width: 100%;
      padding-bottom: 10px;
      border-bottom: 1px solid #E8E8E8;
 }
@@ -44,12 +45,20 @@
      display: flex;
      flex-direction: column;
      padding-top: 17px;
+     max-width: 396px;
+     width: 100%;
  }
- 
- .popup__input_type_email{
- 
- }
- 
+
+.login__input-name-field{
+    margin: 0;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 10px;
+    line-height: 12px;
+    color: #A0A0A0;
+}
+
+
  .login__error{
  
  }
@@ -68,7 +77,8 @@
     border: 0;
     border-radius: 3px;;
     background-color: #FF6838;
-    width: 396px;
+    max-width: 396px;
+    width: 100%;
     font-size: 14px;
     font-weight: 500;
     line-height: 17px;
@@ -101,18 +111,57 @@
     display: flex;
     flex-direction: column;
     padding-top: 17px;
+    max-width: 396px;
+    width: 100%;
  }
 
 .login__input-password-field{
-     margin: 0;
-     font-style: normal;
-     font-weight: 500;
-     font-size: 11px;
-     line-height: 13px;
-     padding-bottom: 5px;
+    margin: 0;
+    font-style: normal;
+    font-weight: 400;
+    font-size: 10px;
+    line-height: 12px;
+    color: #A0A0A0;
 }
 
 .login__input_type_password{
     display: flex;
      flex-direction: column;
+}
+
+@media screen and (max-width: 540px){
+    .login__button-reg{
+        max-width: 260px;
+        width: 100%;
+        font-size: 12px;
+        line-height: 15px;
+    }
+
+    .login__link-signin{
+        font-size: 12px;
+        line-height: 15px;
+    }
+
+    .login__link{
+        font-size: 12px;
+        line-height: 15px;
+    }
+
+    .login__input-email-block{
+        max-width: 260px;
+        width: 100%;
+    }
+
+    .login__input-password-block{
+        max-width: 260px;
+        width: 100%;
+    }
+
+    .login__title{
+        text-align: center;
+    }
+
+    .login__buttons{
+        margin-top: 185px;
+    }
 }
