+++ src/components/Login/Login.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import './Login.css';
 
 function Login() {
     return(
