+++ src/components/Movies/Movies.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
-import SearchForm from '../SearchForm/SearchForm.js';
-import Preloader from '../Preloader/Preloader.js';
-import MoviesCardList from '../MoviesCardList/MoviesCardList.js';
+import SearchForm from '../SearchForm/SearchForm.jsx';
+import Preloader from '../Preloader/Preloader.jsx';
+import MoviesCardList from '../MoviesCardList/MoviesCardList.jsx';
+import './Movies.css';
 
 function Movies () {
     const load = false;
