+++ src/components/Navigation/Navigation.css
@@ -1,21 +1,26 @@
 #menu__toggle {
     opacity: 0;
-  }
-  #menu__toggle:checked + .menu__btn > span {
+}
+
+#menu__toggle:checked + .menu__btn > span {
     transform: rotate(45deg);
-  }
-  #menu__toggle:checked + .menu__btn > span::before {
+}
+
+#menu__toggle:checked + .menu__btn > span::before {
     top: 0;
     transform: rotate(0deg);
-  }
-  #menu__toggle:checked + .menu__btn > span::after {
+}
+
+#menu__toggle:checked + .menu__btn > span::after {
     top: 0;
     transform: rotate(90deg);
-  }
-  #menu__toggle:checked ~ .menu__box {
+}
+
+#menu__toggle:checked ~ .menu__box {
     right: 0 !important;
-  }
-  .menu__btn {
+}
+
+.menu__btn {
     position: fixed;
     top: 30px;
     right: 20px;
@@ -23,26 +28,34 @@
     height: 26px;
     cursor: pointer;
     z-index: 3;
-  }
-  .menu__btn > span,
-  .menu__btn > span::before,
-  .menu__btn > span::after {
+}
+
+.menu__btn:hover{
+    opacity: 0.5;
+}
+
+.menu__btn > span,
+.menu__btn > span::before,
+.menu__btn > span::after {
     display: block;
     position: absolute;
     width: 100%;
     height: 2px;
     background-color: #616161;
     transition-duration: .25s;
-  }
-  .menu__btn > span::before {
+}
+
+.menu__btn > span::before {
     content: '';
     top: -8px;
-  }
-  .menu__btn > span::after {
+}
+
+.menu__btn > span::after {
     content: '';
     top: 8px;
-  }
-  .menu__box {
+}
+
+.menu__box {
     display: block;
     position: fixed;
     top: 0;
@@ -56,9 +69,9 @@
     box-shadow: 2px 2px 6px rgba(0, 0, 0, .4);
     transition-duration: .25s;
     z-index: 2;
-  }
+}
 
-  .menu__item {
+.menu__item {
     display: block;
     padding: 12px 24px;
     color: black;
@@ -70,25 +83,32 @@
     text-align: center;
     text-decoration: none;
     transition-duration: .25s;
-  }
-
-  li:last-of-type{
-    padding-top: 320px;
 }
 
-  .menu__item:hover {
+.menu__box-item:last-of-type{
+    margin: 320px auto 0;
+    border: 1px solid #F9F9F9;
+    border-radius: 20px;
+    background-color: #f9f9f9;
+    max-width: 120px;
+    font-size: 14px;
+    font-weight: 500;
+    line-height: 16px;
+}
+
+.menu__item:hover {
     text-decoration: underline;
     color: black;
-  }
+}
 
-  @media screen and (min-width: 769px) {
+@media screen and (min-width: 769px) {
     .hamburger-menu {
-      display: none;
+        display: none;
     }
-  }
+}
 
-  @media screen and (max-width: 540px) {
+@media screen and (max-width: 540px) {
     .menu__box {
-      width: 315px;
+        width: 315px;
     }
   }
