+++ src/components/Portfolio/Portfolio.css
@@ -1,6 +1,7 @@
 .portfolio{
     margin: 0 auto;
     padding: 0 70px 125px;
+    max-width: 1280px;
 }
 
 .portfolio__title{
