+++ src/components/Portfolio/Portfolio.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Portfolio.css';
 
 function Portfolio () {
     return(
