+++ src/components/Preloader/Preloader.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import './Preloader.css'
+import './Preloader.css';
 
 const Preloader = () => {
     return (
