+++ src/components/Profile/Profile.css
@@ -44,6 +44,7 @@
     justify-content: space-between;
     border-bottom: 1px solid #E8E8E8;
     padding-bottom: 17px;
+    width: 100%;
 }
 
 .profile__input-name-field{
@@ -60,6 +61,7 @@
     align-items: center;
     justify-content: space-between;
     padding-top: 17px;
+    width: 100%;
 }
 
 .popup__input_type_email{
@@ -97,4 +99,11 @@
     font-size: 13px;
     line-height: 16px;
     color: #EE3465;
+}
+
+@media screen and (max-width: 540px){
+    .profile{
+        margin-left: 30px;
+        margin-right: 30px;
+    }
 }
