+++ src/components/Profile/Profile.jsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Profile.css';
 
 function Profile () {
     return(
