+++ src/components/Register/Register.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
+import './Register.css';
 
 function Register() {
     return(
