+++ src/components/Techs/Techs.css
@@ -7,7 +7,7 @@
 .techs__subtitle{
     max-width: 1140px;
     padding: 0 0 23px;
-    margin: 0;
+    margin: 0 auto;
     display: block;
     border-bottom: 1px solid black;
     font-size: 22px;
