+++ src/index.js
@@ -1,8 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './components/App/App.js';
+import App from './components/App/App.jsx';
 import { BrowserRouter } from "react-router-dom"
-import './index.css';
 
 ReactDOM.render(
     <BrowserRouter>
